requirements.txt
cached-property==1.5.2
h5py==3.8.0
joblib==0.17.0
numpy==1.19.4
Pillow==9.4.0
scikit-learn==0.23.2
scipy==1.5.4
threadpoolctl==2.1.0
torch==1.12.1+cu102
torchvision==0.13.1+cu102
torchsummary==1.5.1
tensorboard==2.11.2
tensorflow-gpu==2.11.0
tqdm==4.64.1
natsort==8.2.0
